algorithms/medium/443-string-compression/src/main/java/ru/gadjini/leetcode/_443/Solution.java
package ru.gadjini.leetcode._443;

/**
 * https://leetcode.com/problems/string-compression/
 *
 * "a","a","b","b","c","c","c"
 *
 */
public class Solution {

    public int compress(char[] chars) {
        int i = 0, res = 0;
        while (i < chars.length) {
            int groupLength = 1;
            while (i + groupLength < chars.length && chars[i + groupLength] == chars[i]) {
                groupLength++;
            }
            chars[res++] = chars[i];
            if (groupLength > 1) {
                for (char c : Integer.toString(groupLength).toCharArray()) {
                    chars[res++] = c;
                }
            }
            i += groupLength;
        }
        return res;
    }
}
